# EasySwapBackend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "EasySwap Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/upload/cos-callback": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Handle COS upload callback",
                "parameters": [
                    {
                        "description": "COS callback payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.COSUploadResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/upload/cos-policy": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Get COS upload policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "File type: image/video/audio/document",
                        "name": "file_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/upload/cos-token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Get COS temporary token",
                "parameters": [
                    {
                        "description": "COS token request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types.COSTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.COSTokenResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "types.COSCredentials": {
            "type": "object",
            "properties": {
                "sessionToken": {
                    "type": "string"
                },
                "tmpSecretId": {
                    "type": "string"
                },
                "tmpSecretKey": {
                    "type": "string"
                }
            }
        },
        "types.COSTokenRequest": {
            "type": "object",
            "required": [
                "file_name",
                "file_size",
                "file_type"
            ],
            "properties": {
                "file_name": {
                    "description": "文件名",
                    "type": "string"
                },
                "file_size": {
                    "description": "文件大小（字节）",
                    "type": "integer"
                },
                "file_type": {
                    "description": "文件类型：image, video, audio, document",
                    "type": "string"
                }
            }
        },
        "types.COSTokenResp": {
            "type": "object",
            "properties": {
                "result": {
                    "$ref": "#/definitions/types.COSTokenResponse"
                }
            }
        },
        "types.COSTokenResponse": {
            "type": "object",
            "properties": {
                "allowActions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "allowPrefix": {
                    "type": "string"
                },
                "bucket": {
                    "type": "string"
                },
                "credentials": {
                    "$ref": "#/definitions/types.COSCredentials"
                },
                "expiration": {
                    "type": "string"
                },
                "expiredTime": {
                    "type": "integer"
                },
                "key": {
                    "description": "生成的文件路径",
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "uploadUrl": {
                    "type": "string"
                }
            }
        },
        "types.COSUploadResult": {
            "type": "object",
            "properties": {
                "bucket": {
                    "type": "string"
                },
                "etag": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}